example/requirements.txt
# The django-fluent-contents package has some optional dependencies,
# which don't be installed by default by setup.py
# Instead, this requirements file contains all required elements for the demo:

# Base apps
Django == 1.9.12
Pillow == 3.4.2
pytz == 2016.10

# Requirements for fluent-faq
django-fluent-contents[text,code,markup,oembeditem,picture] >= 1.1.11
django-fluent-utils >= 1.3
django-parler >= 1.7
django-slug-preview >= 1.0.1
django-tag-parser >= 3.0

# Optional requirements
django-categories-i18n >= 1.0
django-taggit >= 0.21.3
django-tinymce >= 2.4.0
taggit-selectize >= 2.3.0

# For other examples:
django-mptt == 0.8.7
django-fluent-pages >= 1.0.1


# These dependencies are all implied by the [text,code,markup,oembeditem,formdesignerlink,picture] construct:
## For code plugin
#Pygments == 2.1.3

## For disquswidgets plugin
#django-disqus>=0.4.1

## For formdesignerlink plugin
#-e git+https://github.com/philomat/django-form-designer.git#egg=django_form_designer-dev

## For markup plugin
#docutils == 0.31.1
#Markdown == 2.6.7
#textile == 2.3.6

## For oembeditem plugin
#micawber == 0.3.3

## For text plugin
#django-wysiwyg == 0.8.0
